[READY] Remove boost where only simple refactoring is needed

This pull request continues with removing boost components with std counterparts. Only simple refactoring is supposed to be done here. Current changes:

- Removes boost/algorithm
- Removes boost/exceptions
- Removes boost/assign
- Removes standard.h in favour of using `size_t`, or `unsigned int`
- Minimises boost/noncopyable, leaving it only in one file

#### Things on a TODO list:

~~Write tests if coverage complains~~
- [x] Bump ycmd core version (Is this needed?)

<!-- Reviewable:start -->
---
This change is [<img src="https://reviewable.io/review_button.svg" height="34" align="absmiddle" alt="Reviewable"/>](https://reviewable.io/reviews/valloric/ycmd/701)
<!-- Reviewable:end -->

+++ cpp/ycm/ClangCompleter/CompilationDatabase.cpp
@@ -17,7 +17,6 @@
 
 #include "CompilationDatabase.h"
 #include "ClangUtils.h"
-#include "standard.h"
 #include "ReleaseGil.h"
 #include "PythonSupport.h"
 
@@ -81,7 +80,7 @@ CompilationInfoForFile CompilationDatabase::GetCompilationInfoForFile(
       compilation_database_,
       path_to_file_string.c_str() ), clang_CompileCommands_dispose );
 
-  uint num_commands = clang_CompileCommands_getSize( commands.get() );
+  size_t num_commands = clang_CompileCommands_getSize( commands.get() );
 
   if ( num_commands < 1 ) {
     return info;
@@ -95,10 +94,10 @@ CompilationInfoForFile CompilationDatabase::GetCompilationInfoForFile(
   info.compiler_working_dir_ = CXStringToString(
                                  clang_CompileCommand_getDirectory( command ) );
 
-  uint num_flags = clang_CompileCommand_getNumArgs( command );
+  size_t num_flags = clang_CompileCommand_getNumArgs( command );
   info.compiler_flags_.reserve( num_flags );
 
-  for ( uint i = 0; i < num_flags; ++i ) {
+  for ( size_t i = 0; i < num_flags; ++i ) {
     info.compiler_flags_.push_back(
       CXStringToString( clang_CompileCommand_getArg( command, i ) ) );
   }
